+++ Makefile.am
@@ -116,7 +116,8 @@ _verbs_files = \
 	prov/verbs/src/verbs_atomic.c \
 	prov/verbs/src/verbs_cm.c \
 	prov/verbs/src/verbs_cq.c \
-	prov/verbs/src/verbs_eq.c
+	prov/verbs/src/verbs_eq.c \
+	prov/verbs/src/verbs_msg_ep.c
 
 if HAVE_VERBS_DL
 pkglib_LTLIBRARIES += libverbs-fi.la
